+++ Trinity.Application/Contracts/IAccountsService.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading.Tasks;
+using Trinity.Application.DTOs.Users;
+
+namespace Trinity.Application.Contracts
+{
+    public interface IAccountsService
+    {
+        Task<UsersOutput> SignUpAccountAsync(UsersInput user);
+        Task<string> SignInAccountAsync(UsersInput user);
+    }
+}
